% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cf.R
\name{make_cf}
\alias{make_cf}
\title{Make Call Frame}
\usage{
make_cf(...)
}
\arguments{
\item{...}{a named list with vectors}
}
\description{
This function will make a call frame given a list of named vectors. This
function will use the purrr::cross_df function to create a dataframe of
every possible call combination determined by the named lists.
}
\details{
if you provide the name of a variable using the | sybmol, then
this variable will be using in the combination then split into two separate
variables afterwards
}
