% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\title{Unpack output from pmap call}
\usage{
unpack(outputs)
}
\arguments{
\item{x}{a list of safely and quielty packaged outputs}
}
\value{
a list of lists such that each element of the parent list contains
all ouput from each function call and each element of the inner list
contains the various outputs (warnings, errors, messages, results, ect.)
from that particular call
}
\description{
When running pmap we have the option to run it safely and quietly. If so,
the output from each call is nested in a list that also contains error
messages, warning messages, echoed output, and what was returned from our
function call.
}
